ajout infos user et nettoyage fichiers
